API_CreateKeyPair to publish the location of private key pair file is saved
o API_CreateKeyPair updates to publish the location where the private
key pair file is saved.
o Updated logs and documentation to provide note to the user to secure
the private key pair file.
o Changed the permissions on the private key pair file to read nly for
the agent user. No other user can access the file.
This is in line with the recommendation that Amazon EC2 provides for the
private key pair file. Currently EC-EC2 plugin does not have specific
use-cases where the generated key pair file is used by other procedures
so we do not want to make any further enhancements to how the file is
handled.